backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
slowapi==0.1.9
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
pytest==8.2.2
httpx==0.27.0
python-multipart==0.0.9
